Introduce useTauriEventHandler
