+++ studio/src/tauri/RuntimeProvider/DefaultRuntime.tsx
@@ -1,0 +1,13 @@
+import { RuntimeContext, type RuntimeProviderProps } from "./RuntimeProvider";
+
+export function DefaultRuntime({ children }: RuntimeProviderProps) {
+  const handleGetApiBaseUrl = () => "";
+
+  return (
+    <RuntimeContext.Provider
+      value={{ type: "unknown", requestApiBaseUrl: handleGetApiBaseUrl }}
+    >
+      {children}
+    </RuntimeContext.Provider>
+  );
+}
