+++ studio/src/tauri/index.ts
@@ -1,1 +1,1 @@
-export * from "./RuntimeProvider";
+export * from "./RuntimeProvider/RuntimeProvider";
